Factor tt::function_info into its own file
