+++ tests/Unit/Pypp/CheckWithRandomValues.hpp
@@ -53,6 +53,7 @@
 #include "Utilities/MakeWithValue.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/FunctionInfo.hpp"
 #include "tests/Unit/Pypp/Pypp.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 #include "tests/Utilities/MakeWithRandomValues.hpp"
